refactor: Refactor pipelines into service
